Rename AggMethods.ts to aggregation.ts
