+++ src/lib/aggregation.ts
